Island: Move construction of AuthenicationFacade
